requirement.txt
tensorflow==2.10.0
keras==2.10.0
numpy==1.23.5
pandas==1.5.3
matplotlib==3.6.3
seaborn==0.12.2
scikit-learn==1.1.3
opencv-python==4.7.0.68
torch==1.13.1
torchvision==0.14.1
Pillow==9.4.0
grad-cam==1.4.6
albumentations==1.3.0
streamlit
